test: unflake some net specs (#18782)
